requirements.txt
beautifulsoup4==4.9.3
gensim==4.1.2
matplotlib==3.3.2
nltk==3.5
 Tokenizer
  sent_tokenize
  word_tokenize
 Corpus
  stopwords
 Stem
  WordNetLemmatizer
numpy==1.19.2
pandas==1.1.3
scikit-learn==0.24.2
 sklearn.feature_extraction.text
  countVectorization
  TfidfTransformer
  TfidfVectorizer
wordcloud==1.8.1
 wordcloud
 stopwords
tqdm==4.64.1